fix: mock vision processor
